esm: experimental wasm modules
This is a first pass at a loader for wasm modules.

It offers support for both named imports as well as deafult.
It may be possible to move some of this logic down to C++, but
I think the current implementation might be good enough. It doesn't
currently use the ESM cache, but AFAICT it isn't currently exposed.

Will dig in more with tests / examples / cache after initial
implementation gets feedback.